refactor: 💡 update tests to consume named export
Previous commit changed public interface to named export. This updates
the tests to work with that.
